requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.0.3
redis==5.0.1
aioredis==2.0.1
httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
ollama==0.1.7
langchain-core==0.3.0
langchain-openai==0.2.0
langgraph==0.2.0
aiosqlite==0.19.0
tenacity==8.2.3
duckduckgo-search==4.1.1
pytest==7.4.4
